+++ NEWS.md
@@ -7,6 +7,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* fixed `set_specification()` and `set_domain_specification()`
+
+* new function `.jsa_jresults()`, `.jread_sap()` and `.jread_workspace()` to  extract the Java object of the results of a SAItem.
+
+* Correction of imports when the workspace contains no SAP.
+
+* `replace` parameter of `save_workspace()` corrected in R (not in Java).
+
 ## [3.2.3] - 2024-07-12