+++ NEWS.md
@@ -12,6 +12,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Rename `.jsa_XXX` in `.jsai_XXX` for SA-Item
 * Rename `.jsap_sa` in `.jsap_sai`
 
+## [3.2.4] - 2025-02-04
+
 ### Fixed
 
 * fixed `set_specification()` and `set_domain_specification()`
@@ -22,7 +24,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * new function `.jsa_jresults()`, `.jread_sap()` and `.jread_workspace()` to  extract the Java object of the results of a SAItem.
 
-
 ## [3.2.3] - 2024-07-12
 
 ### Changed
@@ -31,6 +32,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [3.2.2] - 2024-03-15
 
-[Unreleased]: https://github.com/rjdverse/rjd3workspace/compare/v3.2.3...HEAD
+[Unreleased]: https://github.com/rjdverse/rjd3workspace/compare/v3.2.4...HEAD
+[3.2.4]: https://github.com/rjdverse/rjd3workspace/compare/v3.2.3...v3.2.4
 [3.2.3]: https://github.com/rjdverse/rjd3workspace/compare/v3.2.2...v3.2.3
 [3.2.2]: https://github.com/rjdverse/rjd3workspace/releases/tag/v3.2.2